+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 <!--Releasenotes start-->
 - The 'Reuse new tab' option is now able to reuse the tab even if it is not the currently focused one. It will however not reuse the tab if you have since navigated away from YouTube, to not accidentally overwrite a page you might still need.
 - Added tips and useful information about the extension to the shuffling page opened when using the shuffle button from the popup.
+- Added a tooltip to the 'Shuffle' button on channel and video pages.
 - Fixed an issue with the extension checking an internal URL against a potentially wrong extension ID.
 <!--Releasenotes end-->